refactor: replace libsql crate with turso 0.5.0 and rename plugin to tauri-plugin-turso
- Replace deprecated libsql crate (0.9.29) with turso crate (0.5.0)
- Add turso_core with fts feature to enable Tantivy-based full-text search
- Rename all Rust types (Libsql → Turso), IPC namespace (plugin:libsql → plugin:turso)
- Update guest-js, examples, permissions, docs, and markdown files
- Remove sync/replication/remote modes (local-only for now)
- Add experimental features passthrough (index_method, encryption, triggers, etc.)

BREAKING CHANGE: IPC command namespace changed from plugin:libsql to plugin:turso

+++ src/decode.rs
@@ -1,15 +1,15 @@
-use libsql::{Row, Value};
+use turso::{Row, Value};
 use serde_json::{Number, Value as JsonValue};
 
 use crate::Error;
 
-/// Convert a libsql row column to a JSON value
-pub fn to_json(row: &Row, idx: i32) -> Result<JsonValue, Error> {
+/// Convert a turso row column to a JSON value
+pub fn to_json(row: &Row, idx: usize) -> Result<JsonValue, Error> {
     let value = row.get_value(idx)?;
     value_to_json(value)
 }
 
-/// Convert a libsql Value to a JSON value
+/// Convert a turso Value to a JSON value
 fn value_to_json(value: Value) -> Result<JsonValue, Error> {
     match value {
         Value::Null => Ok(JsonValue::Null),
